fix ws pipe
