+++ test/OpenTelemetry.OpAmp.Client.Tests/OpAmpPipeTests.cs
@@ -7,22 +7,20 @@ using Google.Protobuf;
 using OpAmp.Proto.V1;
 using OpenTelemetry.OpAmp.Client.Internal;
 using OpenTelemetry.OpAmp.Client.Internal.Services.Heartbeat;
-using OpenTelemetry.OpAmp.Client.Listeners;
-using OpenTelemetry.OpAmp.Client.Messages;
 using OpenTelemetry.OpAmp.Client.Settings;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
 using OpenTelemetry.Tests;
 
 namespace OpenTelemetry.OpAmp.Client.Tests;
 
-public class OpAmpPipeTests
+public abstract class OpAmpPipeTests
 {
     [Fact]
     public async Task OpAmpPipe_IsThreadSafe_WhenAppendingConcurrently()
     {
         var taskCount = 20;
 
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -64,7 +62,7 @@ public class OpAmpPipeTests
     [Fact]
     public async Task OpAmpPipe_FlushAsyncCompletes_WhenNoMessagesArePending()
     {
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -77,7 +75,7 @@ public class OpAmpPipeTests
     [Fact]
     public async Task OpAmpPipe_FlushAsyncIsThreadSafe_WhenCalledConcurrentlyWithoutPendingMessages()
     {
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -100,7 +98,7 @@ public class OpAmpPipeTests
     [Fact]
     public async Task OpAmpPipe_FlushesAccumulatedData_WhenCurrentSendFails()
     {
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -124,7 +122,7 @@ public class OpAmpPipeTests
     public async Task OpAmpPipe_LogsSendFailure_WhenCurrentSendFails()
     {
         using var eventListener = new InMemoryEventListener(OpAmpClientEventSource.Log, EventLevel.Verbose);
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -142,102 +140,10 @@ public class OpAmpPipeTests
     }
 
     [Fact]
-    public async Task OpAmpPipe_DoesNotReleaseNewerSend_WhenPublicListenerThrowsDuringHttpResponseProcessing()
-    {
-        var processor = new FrameProcessor();
-        var serverFrame = new ServerToAgent
-        {
-            CustomMessage = new CustomMessage
-            {
-                Data = ByteString.CopyFromUtf8("response"),
-                Type = "Utf8String",
-            },
-        }.ToByteArray();
-
-        using var transport = new MockControlledTransport(
-            () => processor.OnServerFrame(new ReadOnlySequence<byte>(serverFrame)));
-        var settings = new OpAmpClientSettings();
-        using var pipe = new OpAmpPipe(settings, processor, transport);
-
-        processor.Subscribe(new AppendHeartbeatAndThrowListener(pipe));
-
-        AppendIdentification(pipe);
-        await transport.WaitForMessagesAsync(2);
-
-        AppendHeartbeat(pipe);
-
-        await Assert.ThrowsAsync<TimeoutException>(
-            () => transport.WaitForMessagesAsync(3, TimeSpan.FromMilliseconds(250)));
-
-        Assert.Equal(2, transport.Messages.Count);
-    }
-
-    [Fact]
-    public async Task OpAmpPipe_FlushAsyncWaitsForResponseQueuedDataToBeSent()
-    {
-        var processor = new FrameProcessor();
-        var responseWithCustomMessage = new ServerToAgent
-        {
-            CustomMessage = new CustomMessage
-            {
-                Data = ByteString.CopyFromUtf8("response"),
-                Type = "Utf8String",
-            },
-        }.ToByteArray();
-        var emptyResponse = new ServerToAgent().ToByteArray();
-
-        using var transport = new MockControlledTransport();
-        var settings = new OpAmpClientSettings();
-        using var pipe = new OpAmpPipe(settings, processor, transport);
-
-        processor.Subscribe(new AppendHeartbeatListener(pipe));
-
-        AppendIdentification(pipe);
-        await transport.WaitForMessagesAsync(1);
-
-        AppendHeartbeat(pipe);
-        var flushTask = pipe.FlushAsync();
-
-        transport.CompleteNextSend(
-            () => processor.OnServerFrame(new ReadOnlySequence<byte>(responseWithCustomMessage)));
-
-        await transport.WaitForMessagesAsync(2);
-        await Assert.ThrowsAsync<TimeoutException>(
-            () => transport.WaitForMessagesAsync(3, TimeSpan.FromMilliseconds(250)));
-
-        transport.CompleteNextSend(
-            () => processor.OnServerFrame(new ReadOnlySequence<byte>(emptyResponse)));
-
-        await transport.WaitForMessagesAsync(3);
-        Assert.False(flushTask.IsCompleted);
-
-        transport.CompleteNextSend(
-            () => processor.OnServerFrame(new ReadOnlySequence<byte>(emptyResponse)));
-
-#if NET
-        await flushTask.WaitAsync(TimeSpan.FromSeconds(5));
-#else
-        var timeoutTask = Task.Delay(TimeSpan.FromSeconds(5));
-
-        if (await Task.WhenAny(flushTask, timeoutTask).ConfigureAwait(true)
-            == timeoutTask)
-        {
-            throw new TimeoutException("Flush did not complete within 5 seconds.");
-        }
-#endif
-
-        var messages = transport.Messages.ToArray();
-        Assert.Equal([1UL, 2UL, 3UL], messages.Select(m => m.SequenceNum).ToArray());
-        Assert.NotNull(messages[0].AgentDescription);
-        Assert.NotNull(messages[1].Health);
-        Assert.NotNull(messages[2].Health);
-    }
-
-    [Fact]
     public async Task OpAmpPipe_LogsQueueAction_WhenMessageIsAppended()
     {
         using var eventListener = new InMemoryEventListener(OpAmpClientEventSource.Log, EventLevel.Verbose);
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -256,7 +162,7 @@ public class OpAmpPipeTests
     [Fact]
     public async Task OpAmpPipe_StopAsyncWaitsForAccumulatedDataBeforeSendingDisconnect()
     {
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -294,7 +200,7 @@ public class OpAmpPipeTests
     [Fact]
     public async Task OpAmpPipe_StopAsyncCanBeCanceledWhileDrainingQueuedData()
     {
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -317,7 +223,7 @@ public class OpAmpPipeTests
     [Fact]
     public async Task OpAmpPipe_StopAsyncCanBeCanceledWhileWaitingForDisconnectResponse()
     {
-        using var transport = new MockControlledTransport();
+        using var transport = this.GetTransport();
         var settings = new OpAmpClientSettings();
         var processor = new FrameProcessor();
         using var pipe = new OpAmpPipe(settings, processor, transport);
@@ -340,10 +246,12 @@ public class OpAmpPipeTests
         Assert.NotNull(messages[1].AgentDisconnect);
     }
 
-    private static void AppendIdentification(OpAmpPipe pipe)
+    internal abstract MockControlledTransport GetTransport(Action? firstSendCallback = null);
+
+    internal static void AppendIdentification(OpAmpPipe pipe)
         => pipe.AppendMessage(MessageBuilderHelper.AppendIdentification);
 
-    private static void AppendHeartbeat(OpAmpPipe pipe)
+    internal static void AppendHeartbeat(OpAmpPipe pipe)
         => pipe.AppendMessage(MessageBuilderHelper.AppendHeartbeat(CreateHealthReport()));
 
     private static async Task<EventWrittenEventArgs> WaitForEventAsync(InMemoryEventListener eventListener, string eventName, TimeSpan timeout)
@@ -373,33 +281,4 @@ public class OpAmpPipeTests
         IsHealthy = true,
         Status = "OK",
     };
-
-    private sealed class AppendHeartbeatListener : IOpAmpListener<CustomMessageMessage>
-    {
-        private readonly OpAmpPipe pipe;
-
-        public AppendHeartbeatListener(OpAmpPipe pipe)
-        {
-            this.pipe = pipe;
-        }
-
-        public void HandleMessage(CustomMessageMessage message)
-            => AppendHeartbeat(this.pipe);
-    }
-
-    private sealed class AppendHeartbeatAndThrowListener : IOpAmpListener<CustomMessageMessage>
-    {
-        private readonly OpAmpPipe pipe;
-
-        public AppendHeartbeatAndThrowListener(OpAmpPipe pipe)
-        {
-            this.pipe = pipe;
-        }
-
-        public void HandleMessage(CustomMessageMessage message)
-        {
-            AppendHeartbeat(this.pipe);
-            throw new InvalidOperationException("listener failed");
-        }
-    }
 }
